+++ CHANGELOG.md
@@ -13,10 +13,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+## [0.1.0] 2024-01-11
+
+### Added
+
+### Changed
+* Changed `compas_model.elements.Element` class according to the 2nd review including attributes `aabb, obb, id, geometry_points and other`.  
+
+### Removed
+* Removed methods and attributes that are highly dependent on geometry types for example centroid, area, volume and etc.
+
 ## [0.1.0] 2024-01-10
 
 ### Added
-* Added the methods of `compas_model.elements.Element` class.
 
 ### Changed
 * Preparation of the repository for pull requests by removing all element classes.